move classes into entity package
