+++ agr_api/src/main/java/org/alliancegenome/api/controller/GeneController.java
@@ -6,8 +6,8 @@ import org.alliancegenome.api.rest.interfaces.GeneRESTInterface;
 import org.alliancegenome.api.service.DiseaseService;
 import org.alliancegenome.api.service.ExpressionService;
 import org.alliancegenome.api.service.GeneService;
-import org.alliancegenome.api.service.helper.DiseaseRibbonSummary;
-import org.alliancegenome.api.service.helper.ExpressionSummary;
+import org.alliancegenome.api.entity.DiseaseRibbonSummary;
+import org.alliancegenome.api.entity.ExpressionSummary;
 import org.alliancegenome.core.exceptions.RestErrorException;
 import org.alliancegenome.core.exceptions.RestErrorMessage;
 import org.alliancegenome.core.service.JsonResultResponse;
