+++ agr_api/src/main/java/org/alliancegenome/api/rest/interfaces/GeneRESTInterface.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import org.alliancegenome.api.service.helper.DiseaseRibbonSummary;
-import org.alliancegenome.api.service.helper.ExpressionSummary;
+import org.alliancegenome.api.entity.DiseaseRibbonSummary;
+import org.alliancegenome.api.entity.ExpressionSummary;
 import org.alliancegenome.core.service.JsonResultResponse;
 import org.alliancegenome.neo4j.entity.DiseaseAnnotation;
 import org.alliancegenome.neo4j.entity.DiseaseSummary;
