+++ agr_api/src/test/java/org/alliancegenome/api/tests/integration/GeneIT.java
@@ -10,9 +10,9 @@ import org.alliancegenome.api.controller.GeneController;
 import org.alliancegenome.api.controller.GenesController;
 import org.alliancegenome.api.controller.OrthologyController;
 import org.alliancegenome.api.service.GeneService;
-import org.alliancegenome.api.service.helper.ExpressionSummary;
-import org.alliancegenome.api.service.helper.ExpressionSummaryGroup;
-import org.alliancegenome.api.service.helper.ExpressionSummaryGroupTerm;
+import org.alliancegenome.api.entity.ExpressionSummary;
+import org.alliancegenome.api.entity.ExpressionSummaryGroup;
+import org.alliancegenome.api.entity.ExpressionSummaryGroupTerm;
 import org.alliancegenome.core.ExpressionDetail;
 import org.alliancegenome.core.config.ConfigHelper;
 import org.alliancegenome.core.service.JsonResultResponse;
